バグ修正 パース方法変更(作業用RootBox使用)

+++ mp4_parse_test1/mp4_parse_test1/Boxes/SampleEntries.cs
@@ -17,7 +17,7 @@ namespace mp4_parse_test1.Boxes
 			Reserved = new byte[6];
 		}
 
-		public SampleEntry(SampleEntryCode code) : this((BoxType)(uint)code)
+		public SampleEntry(SampleEntryCode code) : this((BoxType)code)
 		{
 		}
 	}
